repo_results_2/HumanEval/33/task_33.py
def sort_third(l: list):
	"""This function takes a list l and returns a list l' such that
	l' is identical to l in the indicies that are not divisible by three, while its values at the indicies that are divisible by three are equal
	to the values of the corresponding indicies of l, but sorted.
	>>> sort_third([1, 2, 3])
	[1, 2, 3]
	>>> sort_third([5, 6, 3, 4, 8, 9, 2])
	[2, 6, 3, 4, 8, 9, 5]
	"""

	# Create a copy of the list
	l_prime = l.copy()

	# Create a list of indices that are divisible by three
	indices = [i for i in range(len(l)) if i % 3 == 0]

	# Create a list of values at these indices
	values = [l[i] for i in indices]

	# Sort the values list
	values.sort()

	# Replace the elements at the indices divisible by three with the sorted values
	for i, value in zip(indices, values):
		l_prime[i] = value

	return l_prime
